Extract the LUA download button into its own component
Move the luaScript query and the download button out of ConfiguratorView
into a self-contained component used by the target and post-build screens.
The query error alert moves with it, so it is now hidden while targets
reload instead of staying visible.

Fixes #158
